+++ Il2CppDumper/ExecutableFormats/NSO.cs
@@ -169,7 +169,7 @@ namespace Il2CppDumper
             {
                 var relaOffset = MapVATR(dynamicSection.First(x => x.d_tag == DT_RELA).d_un);
                 var relaSize = dynamicSection.First(x => x.d_tag == DT_RELASZ).d_un;
-                var relaTable = ReadClassArray<Elf64_Rela>(relaOffset, (long)relaSize / 24L);
+                var relaTable = ReadClassArray<Elf64_Rela>(relaOffset, relaSize / 24L);
                 foreach (var rela in relaTable)
                 {
                     var type = rela.r_info & 0xffffffff;
